+++ src/main/resources/static/js/cart.js
@@ -1,4 +1,7 @@
 const addCartItem = (productId) => {
+    //test
+    alert(productId);
+
     const credentials = localStorage.getItem("credentials");
     if (!credentials) {
         alert("사용자 정보가 없습니다.");
@@ -13,12 +16,23 @@ const addCartItem = (productId) => {
     //         'Authorization': `Basic ${credentials}`
     //     }
     // }).then((response) => {
+
+    // axios
+    //     .post("/cart", productId, {
+    //         headers: {
+    //             Authorization: `Basic ${credentials}`,
+    //         },
+    //     })
+    //     .then((response) => {
+    //         alert("장바구니에 담았습니다.");
+    //     })
+    //     .catch((error) => {
+    //         console.error(error);
+    //     });
+
+
     axios
-        .post("/cart", productId, {
-            headers: {
-                Authorization: `Basic ${credentials}`,
-            },
-        })
+        .post("/cart", productId)
         .then((response) => {
             alert("장바구니에 담았습니다.");
         })
@@ -35,17 +49,8 @@ const removeCartItem = (id) => {
         return;
     }
 
-    // TODO: [2단계] 장바구니 CRUD API에 맞게 변경
-    // axios
-    //     .request({
-    //         url: "",
-    //         headers: {
-    //             Authorization: `Basic ${credentials}`,
-    //         },
-    //     })
-    //     .then((response) => {
     axios
-        .delete("/cart", id, {
+        .delete("/cart/" + id, {
             headers: {
                 Authorization: `Basic ${credentials}`,
             },
