+++ src/main/resources/templates/cart.html
@@ -1,5 +1,6 @@
 <!DOCTYPE html>
 <html xmlns:th="http://www.thymeleaf.org">
+
 <head>
     <meta charset="UTF-8">
     <title>장바구니</title>
@@ -7,63 +8,65 @@
     <link rel="stylesheet" th:href="@{/css/base.css}">
     <link rel="stylesheet" th:href="@{/css/cart.css}">
 </head>
+
 <body>
-<header class="gnb">
-    <nav>
-        <ul class="gnb-group">
-            <li><a href="/">상품목록</a></li>
-            <li><a href="/cart">장바구니</a></li>
-            <li><a href="/settings">설정</a></li>
-            <li class="nav-admin"><a href="/admin">관리자</a></li>
-        </ul>
-    </nav>
-</header>
-<div class="container">
-    <h1>장바구니</h1>
-    <div class="cart-items">
+    <header class="gnb">
+        <nav>
+            <ul class="gnb-group">
+                <li><a href="/">상품목록</a></li>
+                <li><a href="/cart">장바구니</a></li>
+                <li><a href="/settings">설정</a></li>
+                <li class="nav-admin"><a href="/admin">관리자</a></li>
+            </ul>
+        </nav>
+    </header>
+    <div class="container">
+        <h1>장바구니</h1>
+        <div class="cart-items">
+        </div>
     </div>
-</div>
-<script src="https://unpkg.com/axios/dist/axios.min.js"></script>
-<script th:src="@{/js/cart.js}"></script>
-<script>
-    document.addEventListener("DOMContentLoaded", function() {
-        const credentials = localStorage.getItem('credentials');
-        if (!credentials) {
-            alert('사용자 정보가 없습니다.');
-            window.location.href = '/settings';
-            return;
-        }
-
-        axios.request({
-            method: 'post',
-            url: '/cart/list',
-            headers: {
-                'Authorization': `Basic ${credentials}`
+    <script src="https://unpkg.com/axios/dist/axios.min.js"></script>
+    <script th:src="@{/js/cart.js}"></script>
+    <script>
+        document.addEventListener("DOMContentLoaded", function () {
+            const credentials = localStorage.getItem('credentials');
+            if (!credentials) {
+                alert('사용자 정보가 없습니다.');
+                window.location.href = '/settings';
+                return;
             }
-        }).then((response) => {
-            const cartItems = response.data;
 
-            let productElement = '';
-            // TODO: [2단계] 장바구니 CRUD API에 맞게 변경
-            cartItems.forEach(cartItem => {
-                productElement += `
+            axios.request({
+                method: 'post',
+                url: '/cart/list',
+                headers: {
+                    'Authorization': `Basic ${credentials}`
+                }
+            }).then((response) => {
+                const cartItems = response.data;
+
+                let productElement = '';
+                // TODO: [2단계] 장바구니 CRUD API에 맞게 변경
+                cartItems.forEach(cartItem => {
+                    productElement += `
                         <div class="cart-item">
                           <div class="cart-item-info">
-                            <img src="" alt=""/> <!-- 상품 이미지 -->
-                            <div class="cart-item-name">${cartItem.id}</div> <!-- 상품 이름 -->
-                            <div class="cart-item-price"></div> <!-- 상품 가격 -->
+                            <img src=${cartItem.product.image} alt=${cartItem.product.name}/> <!-- 상품 이미지 -->
+                            <div class="cart-item-name">${cartItem.product.name}</div> <!-- 상품 이름 -->
+                            <div class="cart-item-price">${cartItem.product.price}</div> <!-- 상품 가격 -->
                             <button type="submit" id="delete-btn" class="cart-item-delete" onclick="removeCartItem(${cartItem.id})">Delete</button>
                           </div>
                         </div>
                       `;
+                });
+
+                document.querySelector('.cart-items').innerHTML = productElement;
+            }).catch((error) => {
+                console.error(error);
             });
 
-            document.querySelector('.cart-items').innerHTML = productElement;
-        }).catch((error) => {
-            console.error(error);
         });
-
-    });
-</script>
+    </script>
 </body>
+
 </html>
